modify link

+++ src/app/share/auth-partial/auth-partial.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
 import { map } from 'rxjs';
 import { User } from 'src/Models/account/user';
 import { GlobalConstants } from 'src/Models/common/global-constants';
@@ -12,9 +13,12 @@ import { BrowserStorageService } from 'src/services/share/browser-storage.servic
   styleUrls: ['./auth-partial.component.css']
 })
 export class AuthPartialComponent implements OnInit {
-  constructor(private browserStorageService: BrowserStorageService, private router: Router, public accountService: AccountService) { }
+  constructor(private browserStorageService: BrowserStorageService, private toastr: ToastrService,
+    private router: Router, public accountService: AccountService) { }
 
   isLoggedUser: boolean = false;
+  isCustomer: boolean = false;
+  isAdmin: boolean = false;
   panelRoot: string;
 
   adminRole: string = GlobalConstants.AdminRole;
@@ -22,34 +26,63 @@ export class AuthPartialComponent implements OnInit {
 
   ngOnInit(): void {
     this.isLoggedUser = this.accountService.isLogined();
+    this.checkUser();
   }
 
+  // logout() {
+  //   this.browserStorageService.removeLocal("token");
+  //   this.router.navigate(['/'])
+  //   window.location.reload;
+  // }
   logout() {
-    this.browserStorageService.removeLocal("token");
-    this.router.navigate(['/'])
-    window.location.reload;
+    this.accountService.logout().subscribe((response) => {
+      if (response.isSuccess) {
+        this.router.navigate(['/']);
+        this.browserStorageService.removeLocal("token");
+        window.location.reload();
+        return;
+      } else {
+        this.toastr.error(response.message);
+      }
+    });
   }
 
   user: User;
-  goToPanel() {
-    debugger;
+  // goToPanel() {
+  //   debugger;
+  //   this.accountService.getUser().subscribe((response) => {
+  //     if (response.isSuccess) {
+  //       this.user = response.data as User;
+
+  //       let role = this.user.role.filter(x => x.selected == true)[0];
+  //       if (role && role.text == this.adminRole) {
+  //         this.router.navigate(['/Admin/Dashboard']);
+  //         return;
+  //       }
+  //       if (role && role.text == this.customerRole) {
+  //         this.router.navigate(['/CustomerPanel/Profile']);
+  //         return;
+  //       }
+  //       this.router.navigate(['/']);
+
+  //     } else {
+  //       this.router.navigate(['/']);
+  //     }
+  //   });
+  // }
+
+  checkUser() {
     this.accountService.getUser().subscribe((response) => {
       if (response.isSuccess) {
         this.user = response.data as User;
-        
-        let role = this.user.role.filter(x => x.selected == true)[0];
-        if (role && role.text == this.adminRole) {
-          this.router.navigate(['/Admin/Dashboard']);
-          return;
+        let role = this.user.role.filter(x => x.text == this.adminRole && x.selected == true);
+        if (role.length > 0) {
+          this.isAdmin = true;
         }
-        if (role && role.text == this.customerRole) {
-          this.router.navigate(['/CustomerPanel/Profile']);
-          return;
+        role = this.user.role.filter(x => x.text == this.customerRole && x.selected == true);
+        if (role.length > 0) {
+          this.isCustomer = true;
         }
-        this.router.navigate(['/']);
-
-      } else {
-        this.router.navigate(['/']);
       }
     });
   }
